refactor: tool call
