+++ app/tools/__init__.py
@@ -1,0 +1,5 @@
+from .weather_service import get_weather, WeatherArgs
+from .map_service import get_map, MapArgs
+
+__all__ = ['get_weather', 'WeatherArgs',
+           'get_map', 'MapArgs']
